agents/12-lead-scoring-agent/requirements.txt
strands-agents>=0.1.0
boto3>=1.34.0
strands-agents-tools>=0.1.0
pandas>=2.0.0
